@@ -1,32 +1,39 @@
 {
-  "name": "akshith.io",
+  "name": "akshithio",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
     "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
+    "dev": "next dev",
+    "lint": "next lint",
+    "start": "next start"
   },
   "dependencies": {
-    "framer-motion": "^9.0.1",
-    "gpu.js": "^2.16.0",
-    "next": "^14.0.4",
-    "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "react-live-clock": "^6.1.5",
-    "react-text-loop": "^2.3.0",
-    "react-text-loop-next": "^0.0.3"
+    "@t3-oss/env-nextjs": "^0.10.1",
+    "geist": "^1.3.0",
+    "next": "^14.2.1",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "zod": "^3.23.3"
   },
   "devDependencies": {
-    "@types/node": "17.0.38",
-    "@types/react": "18.0.10",
-    "@types/react-dom": "18.0.5",
-    "autoprefixer": "^10.4.7",
-    "eslint": "8.16.0",
-    "eslint-config-next": "12.1.6",
-    "postcss": "^8.4.14",
-    "tailwindcss": "^3.0.24",
-    "typescript": "4.7.2"
-  }
+    "@types/eslint": "^8.56.2",
+    "@types/node": "^20.11.20",
+    "@types/react": "^18.2.57",
+    "@types/react-dom": "^18.2.19",
+    "@typescript-eslint/eslint-plugin": "^7.1.1",
+    "@typescript-eslint/parser": "^7.1.1",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.1.3",
+    "postcss": "^8.4.34",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.4.2"
+  },
+  "ct3aMetadata": {
+    "initVersion": "7.33.1"
+  },
+  "packageManager": "pnpm@9.1.1"
 }
